% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{as.jsonstat}
\alias{as.jsonstat}
\title{Convert JSON-stat object into JSON-stat JSON string}
\usage{
as.jsonstat(x)
}
\arguments{
\item{x}{JSON-stat object}
}
\value{
JSON output
}
\description{
Convert JSON-stat object into JSON-stat JSON string
}
