% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{autounbox}
\alias{autounbox}
\title{Unbox list object}
\usage{
autounbox(.list)
}
\arguments{
\item{.list}{a list contains atomic vectors}
}
\description{
This function marks atomic vectors in given list as a singleton, so that it
will not turn into an 'array' when encoded into JSON.
}
