% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress-plan.R
\name{compress_plan}
\alias{compress_plan}
\title{compress plan}
\usage{
compress_plan(dim_name, role_name = c("time", "geo", "metric",
  "classification", "value"), label = NULL)
}
\arguments{
\item{dim_name}{name of dimension}

\item{role_name}{role of dimension}

\item{label}{label of dimension}
}
\value{
\code{jsonstat.compress.plan} object
}
\description{
This function constructs a plan to compress data frame into jsonstat
}
\examples{
library(jsonstat)

.plan <- compress_plan("place.of.birth", "geo", "Place of Birth")
}
