% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress-plan.R
\name{dimension.jsonstat.compress.plan}
\alias{dimension.jsonstat.compress.plan}
\title{dimension}
\usage{
\method{dimension}{jsonstat.compress.plan}(.plan, dim_name,
  role_name = c("time", "geo", "metric", "classification", "value"),
  label = NULL)
}
\arguments{
\item{.plan}{\code{jsonstat.compress.plan} object}

\item{dim_name}{name of dimension}

\item{role_name}{role of dimension}

\item{label}{label of dimension}
}
\value{
\code{jsonstat.compress.plan} object
}
\description{
This function adds another dimension into compress plan
}
\examples{
library(jsonstat)

.plan <- compress_plan("place.of.birth", "geo", "Place of Birth")
.plan <- dimension(.plan, "age.group", "classification", "Age Group")
}
