% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{dimension}
\alias{dimension}
\title{Create dimension object}
\usage{
dimension(.plan, dim_name, role_name, label)
}
\arguments{
\item{.plan}{\code{jsonstat.compress.plan} object}

\item{dim_name}{name of dimension}

\item{role_name}{role of dimension}

\item{label}{label of dimension}
}
\value{
\code{jsonstat.compress.plan} object
}
\description{
Create dimension object
}
