% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonstat.R
\docType{package}
\name{jsonstat}
\alias{jsonstat}
\alias{jsonstat-package}
\alias{jsonstat-package}
\title{jsonstat: R package for JSON-stat}
\description{
jsonstat provides useful functions to work with JSON-stat format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zedoul/jsonstat}
  \item Report bugs at \url{https://github.com/zedoul/jsonstat/issues}
}

}
\author{
\strong{Maintainer}: Kim Seonghyun \email{zedoul@gmail.com}

}
