% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{toJSON}
\alias{toJSON}
\title{Convert JSON-stat object into JSON-stat JSON string}
\usage{
toJSON(x, auto_unbox)
}
\arguments{
\item{x}{JSON-stat object}

\item{auto_unbox}{this flag marks atomic vectors in given list as a
singleton, so it will not turn into an 'array' when encoded into JSON.
FALSE by default.}
}
\value{
JSON output
}
\description{
Convert JSON-stat object into JSON-stat JSON string
}
